Moved delete API from workflow space to content space
