--- src/main/java/org/craftercms/studio/model/rest/content/DeleteContentRequestBody.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as published by
- * the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.craftercms.studio.model.rest.content;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import javax.validation.Valid;
-import javax.validation.constraints.NotEmpty;
-import java.util.List;
-
-/**
- * Holds data for the deleteContent request
- *
- * @author joseross
- * @since 4.0
- */
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class DeleteContentRequestBody {
-
-    @NotEmpty
-    private String siteId;
-    @NotEmpty
-    private List<@Valid @NotEmpty String> paths;
-    private String submissionComment;
-
-    public String getSiteId() {
-        return siteId;
-    }
-
-    public void setSiteId(String siteId) {
-        this.siteId = siteId;
-    }
-
-    public List<String> getPaths() {
-        return paths;
-    }
-
-    public void setPaths(List<String> paths) {
-        this.paths = paths;
-    }
-
-    public String getSubmissionComment() {
-        return submissionComment;
-    }
-
-    public void setSubmissionComment(String submissionComment) {
-        this.submissionComment = submissionComment;
-    }
-}
