+++ src/main/resources/crafter/studio/extension/rendering-overlay-context.xml
@@ -106,6 +106,7 @@
         <constructor-arg name="siteService" ref="cstudioSiteServiceSimple"/>
         <constructor-arg name="dependencyService" ref="dependencyService"/>
         <constructor-arg name="clipboardService" ref="clipboardService"/>
+        <constructor-arg name="workflowService" ref="studio.workflowService" />
     </bean>
 
     <bean id="pluginController" class="org.craftercms.studio.controller.rest.v2.PluginController">
